requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
dotenv
passlib
python-jose[cryptography]
uuid
pydantic
psycopg2-binary